// apps/backend/src/modules/market-data/dto/market-data.dto.ts
export type WebSocketErrorDto = Readonly<{
  code: string;
  message: string;
  recoverable: boolean;
}>;